[HEAP-8938] Use a `ContentProvider` to auto-initialize Heap for RN+Android.
- Expect (but not require) a `heap.json` file at the root of client
projects that contains the appId. Auto-initialization can also be
disabled using a flag in this file.

- Auto-initialization is nice to have, but is required to fix
sessionization issues on RN+Android builds; the previous approach
(`Heap.setAppId` from Javascript) executed too late, causing the native
Android library to assume that the main activity was in the background
when it was not.

- No action is performed when auto-initialization fails, which can
happen because:

    1. `heap.json` doesn't exist.
    2. `heap.json` doesn't contain the appId.
    3. `heap.json` explicitly disables auto-initialization.

- I plan to update the documentation at
https://docs.heap.io/docs/react-native to include manual initialization
instructions (in `MainActivity`) if a customer really wants to do it
themselves.

- JSON is typically not a great choice for configuration files; it
doesn't support comments, it is a bit verbose, etc. However in this
case, I'd argue that it happens to be the best fit because it is the
only serialization format that both Groovy and Ruby (on the iOS side)
support without having to install any libraries.