Added a script for overwriting Heap.xcframework locally
I relearned the hard way that Cocoapods really doesn't want to load binary xcframework files from on-repo podspecs.

This adds a quick-and-dirty script to the test drivers that can be used to replace the pod-installed framework with a local one, assuming you have the same local repo structure that I do.