[HEAP-42725] Return the same component each time from `withHeapNavigationWrapper`
If `Heap.withHeapNavigationWrapper` is called multiple times, it currently returns a new HOC each time.  This means if the app looks like this:

```
const App = () => {

  ...

  const HeapNavigationContainer =
    Heap.withReactNavigationAutotrack(NavigationContainer);

  return (
    <HeapNavigationContainer independent={true}>
      ...
    </HeapNavigationContainer>
  );
};
```

And the app contains a `useCallback` or similar, which caused `App` to re-evaluate, the entire contents of `HeapNavigationContainer` will be recreated, sending the app back to start.

The general fix is to move `Heap.withHeapNavigationWrapper` out of a function but this has come up enough that a code fix is warranted.

This simply caches the output of `withReactNavigationAutotrack` on the wrapped component so it can be retrieved the next time it is called.