[HEAP-48203] Produce a consistent Cocoapods checksum
[HEAP-48203](https://heapinc.atlassian.net/browse/HEAP-48203)

Reported in #400, the podspec was generating a new checksum on different devices.  The issue is that we were using `path => '.'` instead of `:path => '.'`, which caused the `path` variable to expand as a key.  I've updated to the source repository to match what the modern template does.  It doesn't affect behavior, since `source` is ignored for local podspecs.